[mac] simplify retry success histogram tracking
This commit refactors MAC retry success histogram tracking to simplify
the code and make array indexing cleaner:

- Encapsulate direct and indirect retry success histogram arrays into a
  dedicated `RetryHistogram` struct inheriting from `Clearable`.
- Add `RecordDirectTx()` and `RecordIndirectTx()` helpers that use
  `Min<uint8_t>(aRetryCount, kMax - 1)` to ensure safe indexing and
  accumulate retries exceeding array bounds into the last bucket.
- Simplify `GetDirectRetrySuccessHistogram()` and
  `GetIndirectRetrySuccessHistogram()` implementations using `Min<>`.
- Update API Doxygen comments to document that the last histogram entry
  counts all retries at or above the array limit.

+++ src/ncp/ncp_base_mtd.cpp
@@ -2928,8 +2928,8 @@ template <> otError NcpBase::HandlePropertyGet<SPINEL_PROP_CNTR_MAC_RETRY_HISTOG
     otError         error = OT_ERROR_NONE;
     const uint32_t *histogramDirect;
     const uint32_t *histogramIndirect;
-    uint8_t         histogramDirectEntries;
-    uint8_t         histogramIndirectEntries;
+    uint16_t        histogramDirectEntries;
+    uint16_t        histogramIndirectEntries;
 
     histogramDirect   = otLinkGetTxDirectRetrySuccessHistogram(mInstance, &histogramDirectEntries);
     histogramIndirect = otLinkGetTxIndirectRetrySuccessHistogram(mInstance, &histogramIndirectEntries);
@@ -2939,7 +2939,7 @@ template <> otError NcpBase::HandlePropertyGet<SPINEL_PROP_CNTR_MAC_RETRY_HISTOG
 
     // Encode direct message retries histogram
     SuccessOrExit(error = mEncoder.OpenStruct());
-    for (uint8_t i = 0; i < histogramDirectEntries; i++)
+    for (uint16_t i = 0; i < histogramDirectEntries; i++)
     {
         SuccessOrExit(error = mEncoder.WriteUint32(histogramDirect[i]));
     }
@@ -2947,7 +2947,7 @@ template <> otError NcpBase::HandlePropertyGet<SPINEL_PROP_CNTR_MAC_RETRY_HISTOG
 
     // Encode indirect message retries histogram
     SuccessOrExit(error = mEncoder.OpenStruct());
-    for (uint8_t i = 0; i < histogramIndirectEntries; i++)
+    for (uint16_t i = 0; i < histogramIndirectEntries; i++)
     {
         SuccessOrExit(error = mEncoder.WriteUint32(histogramIndirect[i]));
     }
